+++ node/resolvers/utils/sellerQuotesController.ts
@@ -59,6 +59,30 @@ export default class SellerQuotesController {
     return { organizationName, costCenterName }
   }
 
+  private async getAllChildrenQuotes(parentQuote: string) {
+    const result: Quote[] = []
+
+    const getQuotes = async (page = 1) => {
+      const quotes = await this.ctx.clients.masterdata.searchDocuments<Quote>({
+        dataEntity: QUOTE_DATA_ENTITY,
+        schema: SCHEMA_VERSION,
+        pagination: { page, pageSize: 100 },
+        fields: ['subtotal'],
+        where: `parentQuote=${parentQuote}`,
+      })
+
+      if (quotes.length) {
+        result.push(...quotes)
+
+        await getQuotes(page + 1)
+      }
+    }
+
+    await getQuotes()
+
+    return result
+  }
+
   public async getFullSellerQuote(id: string) {
     const quote = await this.getSellerQuote(id)
     const { organizationName, costCenterName } = await this.getOrganizationData(
@@ -104,5 +128,44 @@ export default class SellerQuotesController {
       data: enrichedQuotes,
       pagination,
     }
+  }
+
+  public async saveSellerQuote(id: string, fields: Partial<Quote>) {
+    const currentQuote = await this.getSellerQuote(id)
+
+    await this.ctx.clients.masterdata.updatePartialDocument({
+      dataEntity: QUOTE_DATA_ENTITY,
+      schema: SCHEMA_VERSION,
+      fields,
+      id,
+    })
+
+    const { subtotal } = currentQuote
+    const subtotalDelta = (fields?.subtotal ?? subtotal) - subtotal
+    const { parentQuote } = fields
+
+    if (!subtotalDelta || !parentQuote) return
+
+    /**
+     * The seller can update the subtotal of your quote changing items
+     * prices. Therefore, it is necessary to update the subtotal of the
+     * parent quote. The call below is asynchronous as there is no need
+     * to stop the flow because of this operation.
+     */
+    this.getAllChildrenQuotes(parentQuote)
+      .then((childrenQuotes) => {
+        const sumSubtotal = childrenQuotes.reduce(
+          (acc, quote) => acc + quote.subtotal,
+          0
+        )
+
+        this.ctx.clients.masterdata.updatePartialDocument({
+          dataEntity: QUOTE_DATA_ENTITY,
+          schema: SCHEMA_VERSION,
+          fields: { subtotal: sumSubtotal },
+          id: parentQuote,
+        })
+      })
+      .catch(() => null)
   }
 }