+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Provides a route for seller get paginated list of quotes at marketplace
+- Provides a route for seller save a quote at marketplace
 
 ## [2.7.0] - 2025-01-09